% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{add.gear.categories}
\alias{add.gear.categories}
\title{Assign gears to categories based on gear and survey}
\usage{
add.gear.categories(data)
}
\arguments{
\item{data}{Data set that includes variables: Gear and Survey}
}
\value{
Data set with added GearCat column.
}
\description{
Assign gears to categories based on gear and survey
}
\details{
See data(gear.categories) for the assignment.
}
