% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.swept.area.R
\name{calc.swept.area}
\alias{calc.swept.area}
\title{Calculate swept area}
\usage{
calc.swept.area(data, plot = FALSE, verbose = TRUE)
}
\arguments{
\item{data}{DATRAS data set (see download.data)}

\item{plot}{Plot some summary graphs. Default: FALSE}

\item{verbose}{Print stuff? Default: TRUE}
}
\value{
Data with updated hh data set with calculated swept area
}
\description{
Calculate swept area
}
\author{
Klaas Sys
}
