% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{correct.species}
\alias{correct.species}
\title{Correct species}
\usage{
correct.species(data)
}
\arguments{
\item{data}{Data set that includes variables: Species and BySpecRecCode}
}
\value{
Data set with corrected Species column
}
\description{
Correct species
}
