% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{correct.surveys}
\alias{correct.surveys}
\title{Correct surveys}
\usage{
correct.surveys(data)
}
\arguments{
\item{data}{Data set that includes variables: Species and BySpecRecCode}
}
\value{
Data set with corrected Species column
}
\description{
Correct surveys
}
