% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est.dist.R
\name{est.dist.one}
\alias{est.dist.one}
\title{Estimated distribution}
\usage{
est.dist.one(specdata, mods = NULL, n.lon = 20, mc.cores = 1)
}
\arguments{
\item{specdata}{Data set of one species}

\item{mods}{NULL default}
}
\value{
List containing hh and hl data sets.
}
\description{
Estimated distribution
}
