% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est.dist.R
\name{est.dist}
\alias{est.dist}
\title{Estimate distribution}
\usage{
est.dist(data, mc.cores = 1)
}
\arguments{
\item{data}{Data set}
}
\value{
List containing hh and hl data sets.
}
\description{
Estimate distribution
}
