% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{est.var.by.area}
\alias{est.var.by.area}
\title{Estimate abudance by ICES Areas}
\usage{
est.var.by.area(fit, grid, mod = 1, cols = 1, by = "Area_27")
}
\arguments{
\item{fit}{fitted object}

\item{grid}{prediction grid}

\item{mod}{Model (Default: 1)}

\item{cols}{ages, if multiple ages were modelled. Default: 1}

\item{by}{area Default: "Area_27"}
}
\value{
data frame with cog in lat and lon for each year
}
\description{
Estimate abudance by ICES Areas
}
