% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishdist-package.R
\docType{package}
\name{fishdist-package}
\alias{fishdist}
\alias{fishdist-package}
\title{fishdist: Estimate distribution maps of fish stocks}
\description{
Estimate distribution maps of fish stocks and abundance indices based on generalized additive models (GAMs).
}
\details{
The only function you're likely to need from roxygen2 is [roxygenize()].
Otherwise refer to the vignettes to see how to format the documentation.
}
\author{
\strong{Maintainer}: Tobias Karl Mildenberger \email{tobm@aqua.dtu.dk} (\href{https://orcid.org/0000-0002-6631-7524}{ORCID}) [copyright holder]

}
\keyword{internal}
