% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotfdist.abun}
\alias{plotfdist.abun}
\title{plot fit}
\usage{
plotfdist.abun(fit, by.area = FALSE, ylab = "Abundance index", y.scale = 1)
}
\arguments{
\item{fit}{fit}

\item{by}{by}
}
\value{
Nothing
}
\description{
plot fit
}
