% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotfdist.diag.spatial}
\alias{plotfdist.diag.spatial}
\title{plot spatial diagnostics}
\usage{
plotfdist.diag.spatial(fit, mod = 1, year = NULL)
}
\arguments{
\item{fit}{fit}

\item{mod}{Select one of the models. Default: 1}
}
\value{
Nothing
}
\description{
plot spatial diagnostics
}
