% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotfdist.dist.cv}
\alias{plotfdist.dist.cv}
\title{plot cv of fit}
\usage{
plotfdist.dist.cv(
  fit,
  mod = NULL,
  year = NULL,
  xlim = NULL,
  ylim = NULL,
  legend = TRUE,
  fixed.scale = TRUE,
  cols = cm.colors(7)[-1],
  breaks = NULL,
  plot.obs = 2
)
}
\arguments{
\item{fit}{fit}

\item{mod}{Select one of the models. Default: NULL}
}
\value{
Nothing
}
\description{
plot cv of fit
}
