% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotfdist.gam.effects.gear}
\alias{plotfdist.gam.effects.gear}
\title{plot gam effects}
\usage{
plotfdist.gam.effects.gear(fit, mod = 1, xlim = NULL, ylim = NULL, CI = 0.95)
}
\arguments{
\item{fit}{fit}

\item{mod}{Select one of the models. Default: 1}
}
\value{
Nothing
}
\description{
plot gam effects
}
