% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est.dist.R
\name{pred.dist.one}
\alias{pred.dist.one}
\title{Predict distribution for given grid or variables}
\usage{
pred.dist.one(
  fit,
  nBoot = 0,
  grid = NULL,
  use.obs.for.pred = TRUE,
  predfix = NULL,
  verbose = TRUE,
  save.grid = FALSE,
  save.data = FALSE,
  ...
)
}
\arguments{
\item{nBoot}{number of bootrstrapping samples}

\item{use.obs.for.pred}{Logical, Only use observed hauls for prediction values? Default: TRUE}

\item{predfix}{Prediction data. Default: NULL}

\item{verbose}{Print stuff? Default: TRUE}

\item{specdata}{Data set of one species}

\item{mods}{NULL default}

\item{use.bathy}{Use bathymetry for distribution. Default: FALSE}

\item{resolution}{Resolution for bathymetry map. Default: 20}

\item{max.dist}{Maximum distance to sample locations for bathymetry map. Default: 3}

\item{var.grid}{Use variable grid by year. Default: TRUE.}
}
\value{
List containing model results
}
\description{
Predict distribution for given grid or variables
}
