% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{pred.statrec}
\alias{pred.statrec}
\title{Predict Statistical rectangle based on latitude and longitude}
\usage{
pred.statrec(data, tol = 1e-05, verbose = TRUE, dbg = 0)
}
\arguments{
\item{data}{Data set that includes variables: lat and lon}
}
\value{
Data set with predicted StatRec if NA.
}
\description{
Predict Statistical rectangle based on latitude and longitude
}
