% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{predict.statrec}
\alias{predict.statrec}
\title{Predict Statistical rectangle based on latitude and longitude}
\usage{
\method{predict}{statrec}(data)
}
\arguments{
\item{data}{Data set that includes variables: StatRec, ShootLat, and ShootLong}
}
\value{
Data set with predicted StatRec if NA.
}
\description{
Predict Statistical rectangle based on latitude and longitude
}
