% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.data.R
\name{prep.data.light}
\alias{prep.data.light}
\title{Prepare data light}
\usage{
prep.data.light(data, aphiaID = NULL)
}
\arguments{
\item{data}{Data set}

\item{aphiaID}{AphiaIDs with species to select. If provided as column of data frame, all additional columns are added. If NULL (default), data is not subsetted for species.}
}
\value{
List containing hh and hl data sets.
}
\description{
Prepare data light
}
