% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{remove.zeros}
\alias{remove.zeros}
\title{Remove zeros from data based on presence-absence model with cutoff}
\usage{
remove.zeros(d, cutat = -3, plot = TRUE, verbose = TRUE, response = "bio")
}
\arguments{
\item{cutat}{default: -3}

\item{plot}{default TRUE}

\item{verbose}{default TRUE}

\item{response}{default "bio"}

\item{data}{data}
}
\description{
Remove zeros from data based on presence-absence model with cutoff
}
\author{
Casper W. Berg
}
