% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{subset.fdist.datras}
\alias{subset.fdist.datras}
\title{Subset}
\usage{
\method{subset}{fdist.datras}(x, subset)
}
\arguments{
\item{x}{object to be subsetted.}

\item{subset}{logical expression indicating elements or rows to keep: missing
values are taken as false.}
}
\description{
Subset
}
