% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.data.R
\name{download.data}
\alias{download.data}
\title{Download all surveys}
\usage{
download.data(
  first.year = 1967,
  last.year = 2020,
  surveys = "all",
  quarters = "all",
  aphiaID = "all",
  datasets = c("HH", "HL"),
  swept.area = TRUE
)
}
\arguments{
\item{first.year}{First year (default: 1967)}

\item{last.year}{Last year (default: 2020)}

\item{surveys}{(default: "all")}

\item{quarters}{(default: "all")}

\item{aphiaID}{Downloads all species anyways, but early subsetting for
species optional and can help with memory issues.}

\item{datasets}{(default: "HH","HL")}

\item{swept.area}{Calculate swept area?}
}
\value{
List containing hh and hl data sets.
}
\description{
Download all surveys
}
