% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est.dist.R
\name{est.dist.one}
\alias{est.dist.one}
\title{Estimated distribution}
\usage{
est.dist.one(
  specdata,
  mods = NULL,
  n.lon = 20,
  nBoot = 0,
  use.bathy = FALSE,
  resolution = 20,
  verbose = TRUE
)
}
\arguments{
\item{specdata}{Data set of one species}

\item{mods}{NULL default}

\item{nBoot}{number of bootrstrapping samples}

\item{use.bathy}{Use bathymetry for distribution. Default: FALSE}

\item{resolution}{Resolution for bathymetry map}

\item{verbose}{Print stuff? Default: TRUE}
}
\value{
List containing model results
}
\description{
Estimated distribution
}
