% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{list.recom.models}
\alias{list.recom.models}
\title{List all model recommended structures}
\usage{
list.recom.models(specdata, use.toy = TRUE, use.swept.area = TRUE)
}
\arguments{
\item{specdata}{Species data}

\item{use.toy}{Use time of year? (Caused problems for some species)}

\item{use.swept.area}{Use swept area? (Might not be available)}
}
\value{
List with all model structures
}
\description{
List all model recommended structures
}
