% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotfdist.dist}
\alias{plotfdist.dist}
\title{plot fit}
\usage{
plotfdist.dist(fit, mod = NULL)
}
\arguments{
\item{fit}{fit}

\item{mod}{Select one of the models. Default: NULL}
}
\value{
Nothing
}
\description{
plot fit
}
