% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.species.R
\name{prep.species}
\alias{prep.species}
\title{Prepare species}
\usage{
prep.species(
  data,
  aphiaID,
  use.gear.cat = TRUE,
  remove.fragmented.years = TRUE,
  min.gears = 1,
  min.surveys = 1,
  min.hauls = 1,
  min.ship.gear = 1
)
}
\arguments{
\item{data}{Data set}

\item{aphiaID}{required (single species or vector of species)}

\item{use.gear.cat}{= TRUE}

\item{remove.fragmented.years}{= TRUE}

\item{min.gears}{= 1}

\item{min.surveys}{= 1}

\item{min.hauls}{= 1}

\item{min.ship.gear}{= 1}
}
\value{
List containing hh and hl data sets.
}
\description{
Prepare species
}
